+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Keep focus on the text input after sending a message through the overlay.
 
+* Unfocus and fade out the overlay when loading into a game.
+
 ## 4.0.0-beta.3
 
 * Add `Drops` and `Shops` location groups.